refactor(server): sort models
