+++ Server/Models/Auth/UserRegistrationRequestDto.cs
@@ -1,6 +1,7 @@
 ﻿using Server.Entities;
+using Server.Models.Extensions;
 
-namespace Server.Models;
+namespace Server.Models.Auth;
 
 public static class UserRegistrationDto
 {
