+++ Server/Models/Collection/CollectionCreationDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Server.Models;
+﻿namespace Server.Models.Collection;
 
 public static class CollectionCreationDto
 {
