+++ Server/Models/CollectionItemSpecialStatus/CollectionItemSpecialStatusUpdateDto.cs
@@ -1,4 +1,6 @@
-﻿namespace Server.Models;
+﻿using Server.Models.Extensions;
+
+namespace Server.Models.CollectionItemSpecialStatus;
 
 public static class CollectionItemSpecialStatusUpdateDto
 {
