+++ Server/Models/Country/CountryCreationDto.cs
@@ -1,4 +1,6 @@
-﻿namespace Server.Models;
+﻿using Server.Models.Extensions;
+
+namespace Server.Models.Country;
 
 public static class CountryCreationDto
 {
