+++ Server/Services/CatalogItemService.cs
@@ -1,6 +1,6 @@
 ﻿using Server.Data;
 using Server.Entities;
-using Server.Models;
+using Server.Models.CatalogItem;
 
 namespace Server.Services;
